add delete message view

+++ home/views.py
@@ -96,6 +96,7 @@ def home(request):
 def room(request, pk):
     room = Room.objects.get(id=pk)
     room_messages = room.message_set.all().order_by('-created')
+    participants = room.participants.all()
 
     if request.method == "POST":
         message = Message.objects.create(
@@ -103,11 +104,13 @@ def room(request, pk):
             room = room,
             body = request.POST.get('body')
         )
+        room.participants.add(request.user)
         return redirect('home:room', pk=room.id)
 
     return render(request, 'home/room.html', {
         'room':room,
-        'room_messages':room_messages
+        'room_messages':room_messages,
+        'participants': participants
     }
     )
 
@@ -151,9 +154,29 @@ def UpdateRoom(request, pk):
 def DeleteRoom(request, pk):
     template_name = 'home/delete.html'
     room = Room.objects.get(id=pk) 
+    if request.user != room.host:
+        return HttpResponse("You are not allowed here!")
+
     if request.method == 'POST':
         room.delete()
         return redirect('home:home')
     return render(request, template_name, {'obj':room})
 
+
+
+
+@login_required(login_url='home:login')
+def DeleteMessage(request, pk):
+    template_name = 'home/delete.html'
+    message = Message.objects.get(id=pk) 
+
+
+    if request.user != message.user:
+        return HttpResponse("You are not allowed here!")
+
+
+    if request.method == 'POST':
+        message.delete()
+        return redirect('home:home')
+    return render(request, template_name, {'obj':message})
